Inadvertant debugging on satellite project
